+++ src/check/index.ts
@@ -1,0 +1,2 @@
+export * from "./isObject";
+export * from "./isString";
