+++ src/check/isObject.ts
