+++ src/check/isString.ts
