+++ src/objects/index.ts
@@ -1,2 +1,1 @@
 export * from "./hasKey";
-export * from "./isObject";
