+++ src/strings/index.ts
@@ -1,2 +1,1 @@
 export * from "./capitalize";
-export * from "./isString";
